+++ application-context/realworld-spring-boot/src/main/java/com/github/hhjin015/repository/TodoRepository.java
@@ -1,14 +1,15 @@
 package com.github.hhjin015.repository;
 
 import com.github.hhjin015.Todo;
+
 import java.util.List;
 
 public interface TodoRepository {
     void save(Todo todo);
 
-    Todo findById(int id);
+    Todo findById(long id);
 
     List<Todo> findAll();
 
-    Todo deleteById(int id);
+    Todo deleteById(long id);
 }
